requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==4.2.13
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-crispy-forms==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Email==1.4.4
Flask-Mail==0.9.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.4
pdfkit==1.0.0
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
reportlab==4.0.6
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
